add regex
Typos typed in the custom DNS field could cause issues. The program will check the validity of a DNS address with a regex function to help mitigate this.

If the DNS address is not recognizable (e.g. 192.q18.1.w or 2,2,2,2 instead of 192.168.1.1), the program will prompt the user with a yes or no prompt (default no) to confirm if they really want the said address.